+++ src/lib/utils/options/readers/typedoc.ts
@@ -33,7 +33,7 @@ export class TypeDocReader implements OptionsReader {
         if (!file) {
             if (container.isSet("options")) {
                 logger.error(
-                    `The options file could not be found with the given path ${path}`
+                    `The options file ${nicePath(path)} does not exist.`
                 );
             }
             return;